csi_plugin.py
from concurrent import futures  
import grpc  
import logging  
from csi_pb2 import (  
    GetPluginInfoResponse,  
    GetPluginCapabilitiesResponse,  
    PluginCapability,  
    ProbeResponse,  
    CreateVolumeResponse,  
    Volume,  
    DeleteVolumeResponse,  
)  
from csi_pb2_grpc import IdentityServicer, ControllerServicer, NodeServicer, add_IdentityServicer_to_server, add_ControllerServicer_to_server, add_NodeServicer_to_server  

# Configure logging to include time, filename, and log message
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s %(filename)s: %(message)s',
    datefmt='%Y-%m-%d %H:%M:%S'
)

class IdentityService(IdentityServicer):  
    def GetPluginInfo(self, request, context):  
        logging.info("GetPluginInfo called")
        return GetPluginInfoResponse(name="my-csi-plugin", vendor_version="v0.1")  

    def GetPluginCapabilities(self, request, context):  
        logging.info("GetPluginCapabilities called")
        return GetPluginCapabilitiesResponse(  
            capabilities=[PluginCapability(service=PluginCapability.Service(type=PluginCapability.Service.CONTROLLER_SERVICE))]  
        )  

    def Probe(self, request, context):  
        logging.info("Probe called")
        return ProbeResponse()  

class ControllerService(ControllerServicer):  
    def CreateVolume(self, request, context):  
        logging.info("CreateVolume called")
        # Implement volume creation logic  
        return CreateVolumeResponse(  
            volume=Volume(volume_id="my-volume-id", capacity_bytes=request.capacity_range.required_bytes)  
        )  

    def DeleteVolume(self, request, context):  
        logging.info("DeleteVolume called")
        # Implement volume deletion logic  
        return DeleteVolumeResponse()  

class NodeService(NodeServicer):  
    def NodeStageVolume(self, request, context):  
        logging.info("NodeStageVolume called")
        # Implement node stage volume logic  
        return super().NodeStageVolume(request, context)  

    def NodeUnstageVolume(self, request, context):  
        logging.info("NodeUnstageVolume called")
        # Implement node unstage volume logic  
        return super().NodeUnstageVolume(request, context)  

def serve():  
    server = grpc.server(futures.ThreadPoolExecutor(max_workers=10))  
    add_IdentityServicer_to_server(IdentityService(), server)  
    add_ControllerServicer_to_server(ControllerService(), server)  
    add_NodeServicer_to_server(NodeService(), server)  
    server.add_insecure_port("[::]:50051")  
    logging.info("Starting CSI plugin...")
    server.start()  
    server.wait_for_termination()  

if __name__ == "__main__":  
    logging.basicConfig()  
    serve()